Update rules.py
This change is just a quick initial test. No point in creating all classes for each rule until I am sure that this structure works as it is. I have used the primitive type " None ". In PyCharm my expected output matched my actual so I was happy with these results.